Warning squashing. 64-bit improvements. Fixed few bugs. Improved the typing of functions that take a naming method of variable.
git-svn-id: http://svn.cri.ensmp.fr/svn/pips/trunk@12872 32a0c1e6-9308-0410-80dd-f9d2257a62c7

+++ src/Libs/c_syntax/c_parser.c
@@ -125,7 +125,7 @@ void reset_keyword_typedef_table()
 intptr_t
 is_c_keyword_typedef(char * s)
 {
-  intptr_t i = hash_get(keyword_typedef_table,s);
+  intptr_t i = (intptr_t) hash_get(keyword_typedef_table,s);
   return ((char *) i == HASH_UNDEFINED_VALUE) ? 0: i;
 }
